% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_metrics.R
\name{cran_metrics}
\alias{cran_metrics}
\title{Get CRAN-related metrics for a package.}
\usage{
cran_metrics(package_name, forget = FALSE)
}
\arguments{
\item{package_name}{name of CRAN package - case-sensitive.}
}
\value{
a tibble of package information.
}
\description{
Searches CRAN database to pull information like dependencies, reverse dependencies, etc.
}
\examples{

cran_metrics("dplyr")
}
