% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cran.R
\name{get_cran}
\alias{get_cran}
\title{Get CRAN Packages}
\usage{
get_cran(forget = FALSE)
}
\arguments{
\item{forget}{logical, default \code{FALSE}. Reset the CRAN package cache.}
}
\description{
Returns a tibble containing the current CRAN pacakges.
For efficiency reasons, the tibble is cached, so future calls will just return
the original tibble.
}
\examples{
cran <- get_cran()
}
