% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackOverflow.R
\name{get_hits}
\alias{get_hits}
\title{Get number of search results on StackOverflow}
\usage{
get_hits(pkg, tag = "r", pagesize = 100, num_pages = 200, ...)
}
\arguments{
\item{pkg}{Which package are you intersted in?}

\item{tag}{StackOverflow tag.}

\item{pagesize}{Size of each page to extract (max 100).}

\item{num_pages}{Number of pages to extract.}

\item{...}{Passed on to \code{stack_search}.}
}
\description{
Get number of search results on StackOverflow
}
