% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackOverflow.R
\name{get_pkgs}
\alias{get_pkgs}
\title{Get a list of CRAN packages associated with a topic}
\usage{
get_pkgs(topics, tag = "r", pagesize = 100, num_pages = 200,
  repos = "https://cran.rstudio.com", ...)
}
\arguments{
\item{topics}{What are you trying to do?}

\item{tag}{StackOverflow tag.}

\item{pagesize}{Size of each page to extract (max 100).}

\item{num_pages}{Number of pages to extract.}

\item{repos}{Vector of repository URLs passed on to \code{available.packages}.}

\item{...}{Passed on to \code{stack_search}.}
}
\value{
A vector of packages mentioned in StackOverflow questions on the \code{topic}.
}
\description{
Get a list of CRAN packages associated with a topic
}
\examples{
\dontrun{
get_pkgs(topics = "table", tag = "r", pagesize = 100, num_pages = 2)
get_pkgs(topics = c("table", "memory"), tag = "r", pagesize = 100, num_pages = 2)
}
}
