% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGitHub.R
\name{getGitHub}
\alias{getGitHub}
\title{Get GitHub URLs for a set of package names (if they exist).}
\usage{
getGitHub(packages, forget = FALSE)
}
\arguments{
\item{packages}{character vector of CRAN package names - case-sensitive.}
}
\value{
a vector of URLs.
}
\description{
Looks at both Package URL and BugReports URL, returns GitHub URL if present in either.
}
\examples{

getGitHub("dplyr")
}
