% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_table.R
\name{metrics_table}
\alias{metrics_table}
\title{Formatted table of package metrics}
\usage{
metrics_table(pkg_df)
}
\arguments{
\item{pkg_df}{A data.frame of package metrics as returned by \code{ackage_list_metrics}.}
}
\description{
Produce a formatted table of package metrics.
}
\examples{

pkgs <- package_list_metrics(list("dplyr", "tidyr", "ggplot2"))
metrics_table(pkgs)
}
\seealso{
\code{package_list_metrics}
}
