% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_list_metrics.R
\name{package_list_metrics}
\alias{package_list_metrics}
\title{Get metrics for a list of packages.}
\usage{
package_list_metrics(package_names_list, forget = FALSE)
}
\arguments{
\item{package_names_list}{list of CRAN packages - case-sensitive.}

\item{forget}{logical, default \code{FALSE}. Reset the CRAN package cache.}
}
\value{
a tibble of package information
}
\description{
wrapper for combine_metrics for list of packages
}
\examples{

package_list_metrics(list("dplyr", "tidyr", "ggplot2"))
}
