% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagemetrics.R
\docType{package}
\name{packagemetrics}
\alias{packagemetrics}
\alias{packagemetrics-package}
\title{janitor}
\description{
packagemetrics gathers information about a package from CRAN, StackOverflow, and GitHub (if applicable).
}
\keyword{internal}
