% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackOverflow.R
\name{so_metrics}
\alias{so_metrics}
\title{Get StackOverflow-related metrics for a package.}
\usage{
so_metrics(package_name)
}
\arguments{
\item{package_name}{name of package - case sensitive.}
}
\value{
A data.frame of package information.
}
\description{
Get StackOverflow-related metrics for a package.
}
\examples{
\dontrun{
so_metrics("dplyr")
}
}
