% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tables}
\alias{tables}
\title{Table of Packages}
\format{An object of class \code{character} of length 27.}
\usage{
tables
}
\description{
@format A character vector of packages
}
\keyword{datasets}
