% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_metrics.R
\name{combine_metrics}
\alias{combine_metrics}
\title{Combines metrics on a package from CRAN and Github}
\usage{
combine_metrics(package_name, includeSO = FALSE)
}
\arguments{
\item{package_name}{name of CRAN package - case-sensitive.}

\item{includeSO}{Logical. Include metrics from StackOverflow?}
}
\value{
a tibble of package information.
}
\description{
Combines metrics on a package from CRAN and Github
}
\examples{

combine_metrics("dplyr")

}
